chore(coin-evm): relocate `operations.ts` to `ledger-live-common`
