+++ .changeset/relocate-evm-operation-helpers.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/coin-evm": patch
+"@ledgerhq/live-common": patch
+---
+
+Relocate EVM operation helpers (`isEditableOperation`, `isStuckOperation`, `getStuckAccountAndOperation`) from `@ledgerhq/coin-evm/operation` to `families/evm/editTransaction/` in `ledger-live-common`
