+++ libs/coin-modules/coin-evm/package.json
@@ -95,7 +95,6 @@
     "jest-expect-message": "^1.1.3",
     "msw": "catalog:",
     "oxfmt": "catalog:",
-    "oxlint": "catalog:",
-    "prando": "^6.0.1"
+    "oxlint": "catalog:"
   }
 }
